collection/collector/requirements.txt
pyzmq
pandas
requests
#Websockets 6.0 because 7.0 has internal 1006 error crash no reason.
websockets==6.0
signalr-client-aio
simplejson
SQLAlchemy
psycopg2